+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2022-05-16
+### Added
+- `kotlinx.serialization` support with predefined serializers.
+- Equality operator for `Constraint`.
+
 ## [1.2.1] - 2022-01-24
 ### Fixed
 - There was a case where the `ConstraintFormatException` did not contain the invalid Constraint's details.